Move resizing width and background logic to `SongSelectComponentsTestScene`
